+++ src/main/java/cafe/controller/handler/Handler.java
@@ -1,4 +1,4 @@
-package cafe.handler;
+package cafe.controller.handler;
 
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServletRequest;
