+++ src/main/java/cafe/controller/handler/articles/ArticleHandler.java
@@ -1,6 +1,6 @@
-package cafe.handler.articles;
+package cafe.controller.handler.articles;
 
-import cafe.handler.Handler;
+import cafe.controller.handler.Handler;
 import cafe.service.ArticleService;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletContext;
@@ -10,26 +10,22 @@ import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
 
-public class ArticleCreateHandler implements Handler {
+public class ArticleHandler implements Handler {
     private final ArticleService articleService;
 
-    public ArticleCreateHandler(ServletContext context) {
+    public ArticleHandler(ServletContext context) {
         articleService = (ArticleService) context.getAttribute("articleService");
     }
 
     @Override
     public void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
-        RequestDispatcher dispatcher = req.getRequestDispatcher("/WEB-INF/article/form.jsp");
+        req.setAttribute("article", articleService.find(req.getRequestURI()));
+        RequestDispatcher dispatcher = req.getRequestDispatcher("/WEB-INF/article/show.jsp");
         dispatcher.forward(req, resp);
     }
 
     @Override
-    public void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
-        String writer = req.getParameter("writer");
-        String title = req.getParameter("title");
-        String contents = req.getParameter("contents");
-
-        articleService.save(writer, title, contents);
-        resp.sendRedirect("/");
+    public void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
+        doGet(req, resp);
     }
 }
