+++ src/main/java/cafe/controller/handler/articles/ArticleListHandler.java
@@ -1,6 +1,6 @@
-package cafe.handler.articles;
+package cafe.controller.handler.articles;
 
-import cafe.handler.Handler;
+import cafe.controller.handler.Handler;
 import cafe.service.ArticleService;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletContext;
@@ -10,22 +10,17 @@ import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
 
-public class ArticleHandler implements Handler {
+public class ArticleListHandler implements Handler {
     private final ArticleService articleService;
 
-    public ArticleHandler(ServletContext context) {
+    public ArticleListHandler(ServletContext context) {
         articleService = (ArticleService) context.getAttribute("articleService");
     }
 
     @Override
     public void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
-        req.setAttribute("article", articleService.find(req.getRequestURI()));
-        RequestDispatcher dispatcher = req.getRequestDispatcher("/WEB-INF/article/show.jsp");
+        req.setAttribute("articles", articleService.findAll());
+        RequestDispatcher dispatcher = req.getRequestDispatcher("/WEB-INF/article/list.jsp");
         dispatcher.forward(req, resp);
-    }
-
-    @Override
-    public void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
-        doGet(req, resp);
     }
 }
