+++ src/main/java/cafe/controller/handler/users/UserInfoEditHandler.java
@@ -1,6 +1,6 @@
-package cafe.handler.users;
+package cafe.controller.handler.users;
 
-import cafe.handler.Handler;
+import cafe.controller.handler.Handler;
 import cafe.service.UserService;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletContext;
@@ -10,27 +10,29 @@ import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
 
-public class UserSignUpHandler implements Handler {
+public class UserInfoEditHandler implements Handler {
     private final UserService userService;
 
-    public UserSignUpHandler(ServletContext context) {
+    public UserInfoEditHandler(ServletContext context) {
         userService = (UserService) context.getAttribute("userService");
     }
 
     @Override
     public void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
-        RequestDispatcher dispatcher = req.getRequestDispatcher("/WEB-INF/user/form.jsp");
+        req.setAttribute("user", userService.find(req.getRequestURI()));
+        RequestDispatcher dispatcher = req.getRequestDispatcher("/WEB-INF/user/update_form.jsp");
         dispatcher.forward(req, resp);
     }
 
     @Override
     public void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
-        String id = req.getParameter("userid");
+        String uri = req.getRequestURI();
         String name = req.getParameter("name");
         String password = req.getParameter("password");
         String email = req.getParameter("email");
+        String beforePassword = req.getParameter("before-password");
 
-        userService.save(id, name, password, email);
+        userService.update(uri, name, password, email, beforePassword);
         resp.sendRedirect("/users");
     }
 }
