+++ src/main/java/cafe/controller/handler/users/UserInfoListHandler.java
@@ -1,6 +1,6 @@
-package cafe.handler.users;
+package cafe.controller.handler.users;
 
-import cafe.handler.Handler;
+import cafe.controller.handler.Handler;
 import cafe.service.UserService;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletContext;
