+++ src/main/java/cafe/controller/handler/users/UserSignUpHandler.java
@@ -1,6 +1,6 @@
-package cafe.handler.users;
+package cafe.controller.handler.users;
 
-import cafe.handler.Handler;
+import cafe.controller.handler.Handler;
 import cafe.service.UserService;
 import jakarta.servlet.RequestDispatcher;
 import jakarta.servlet.ServletContext;
@@ -10,29 +10,27 @@ import jakarta.servlet.http.HttpServletResponse;
 
 import java.io.IOException;
 
-public class UserInfoEditHandler implements Handler {
+public class UserSignUpHandler implements Handler {
     private final UserService userService;
 
-    public UserInfoEditHandler(ServletContext context) {
+    public UserSignUpHandler(ServletContext context) {
         userService = (UserService) context.getAttribute("userService");
     }
 
     @Override
     public void doGet(HttpServletRequest req, HttpServletResponse resp) throws IOException, ServletException {
-        req.setAttribute("user", userService.find(req.getRequestURI()));
-        RequestDispatcher dispatcher = req.getRequestDispatcher("/WEB-INF/user/update_form.jsp");
+        RequestDispatcher dispatcher = req.getRequestDispatcher("/WEB-INF/user/form.jsp");
         dispatcher.forward(req, resp);
     }
 
     @Override
     public void doPost(HttpServletRequest req, HttpServletResponse resp) throws IOException {
-        String uri = req.getRequestURI();
+        String id = req.getParameter("userid");
         String name = req.getParameter("name");
         String password = req.getParameter("password");
         String email = req.getParameter("email");
-        String beforePassword = req.getParameter("before-password");
 
-        userService.update(uri, name, password, email, beforePassword);
+        userService.save(id, name, password, email);
         resp.sendRedirect("/users");
     }
 }
