+++ src/main/java/cafe/domain/util/SQLGenerator.java
@@ -1,4 +1,4 @@
-package cafe.domain.sql;
+package cafe.domain.util;
 
 import java.lang.reflect.Field;
 
@@ -15,14 +15,20 @@ public class SQLGenerator {
     }
 
     public String generateInsertSQL(String className, Field[] fields) {
+        if (className == null || fields == null) {
+            throw new IllegalArgumentException("Invalid arguments");
+        }
+
         StringBuilder stringBuilder = new StringBuilder("INSERT INTO `").append(className).append("s` (");
         stringBuilder.append("`id`, ");
         for (Field field : fields) {
+            if (field.getName().equals("this$0")) continue;
             stringBuilder.append("`").append(field.getName()).append("`, ");
         }
         stringBuilder.delete(stringBuilder.length() - 2, stringBuilder.length());
         stringBuilder.append(") VALUES (?, ");
         for (Field field : fields) {
+            if (field.getName().equals("this$0")) continue;
             stringBuilder.append("?, ");
         }
         stringBuilder.delete(stringBuilder.length() - 2, stringBuilder.length());
@@ -31,16 +37,27 @@ public class SQLGenerator {
     }
 
     public String generateSelectByIdSQL(String className) {
+        if (className == null) {
+            throw new IllegalArgumentException("Invalid arguments");
+        }
         return "SELECT * FROM `" + className + "s` WHERE `id` = ?;";
     }
 
     public String generateSelectAllSQL(String className) {
+        if (className == null) {
+            throw new IllegalArgumentException("Invalid arguments");
+        }
         return "SELECT * FROM `" + className + "s`;";
     }
 
     public String generateUpdateSQL(String className, Field[] fields) {
+        if (className == null || fields == null) {
+            throw new IllegalArgumentException("Invalid arguments");
+        }
+
         StringBuilder stringBuilder = new StringBuilder("UPDATE `").append(className).append("s` SET ");
         for (Field field : fields) {
+            if (field.getName().equals("this$0")) continue;
             stringBuilder.append("`").append(field.getName()).append("` = ?, ");
         }
         stringBuilder.delete(stringBuilder.length() - 2, stringBuilder.length());
